requirements.txt
Flask
gunicorn
python-dotenv
flask-cors
openai
google-generativeai
google-cloud-aiplatform
google-cloud-logging
google-cloud-storage
twilio
pywebpush
py-vapid
psycopg2-binary
pgvector
apscheduler
cryptography
Pillow
qrcode
numpy
tiktoken
python-dateutil
